Move executor to common
