+++ common/CMakeLists.txt
@@ -8,6 +8,8 @@ file(GLOB COMMON_SRCS
         "${CMAKE_CURRENT_LIST_DIR}/src/memory/*.c"
         "${CMAKE_CURRENT_LIST_DIR}/src/stack/*.c"
         "${CMAKE_CURRENT_LIST_DIR}/src/utils/*.c"
+        "${CMAKE_CURRENT_LIST_DIR}/src/execution_engine/*.c"
+        "${CMAKE_CURRENT_LIST_DIR}/src/native/*.c"
 )
 
 add_library(common ${COMMON_SRCS})
