+++ common/src/native/stdlib_native.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-#include "native.h"
+#include "native/native.h"
 
 void loadNativeStdLib() {
     loadSharedLibrary("libstdlib_native.so");
